requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
ecdsa==0.18.0
fastai==2.7.13
fastapi==0.108.0
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
h11==0.14.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
mangum==0.17.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pillow==10.2.0
preshed==3.0.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pypyodbc==1.3.6
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.32.0.post1
sympy==1.12
thinc==8.2.2
threadpoolctl==3.2.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
wasabi==1.1.2
weasel==0.3.4
